+++ frontend/src/widgets/fixture/fixture-widget.tsx
@@ -5,7 +5,7 @@ import { useMasterDataMaps, useApiStateEntry } from '../../hooks/api'
 import { setFixtureState } from '../../api'
 import { memoInProduction } from '../../util/development'
 import { iconLight } from '../../ui/icons'
-import { useEvent } from '../../hooks/performance'
+import { useEvent, useShallowEqualMemo } from '../../hooks/performance'
 import { WidgetPassthrough } from '../../ui/containers/widget'
 
 import { FixtureStateWidget } from './fixture-state-widget'
@@ -22,13 +22,14 @@ export function getFixtureName(
 
 export interface FixtureWidgetProps extends WidgetPassthrough {
   fixture: Fixture
+  mapping?: string[]
 }
 
 /**
  * Widget to display a fixture.
  */
 export const FixtureWidget = memoInProduction(
-  ({ fixture, ...passThrough }: FixtureWidgetProps) => {
+  ({ fixture, mapping, ...passThrough }: FixtureWidgetProps) => {
     const { fixtureTypes } = useMasterDataMaps()
     const fixtureType = fixtureTypes.get(fixture.type)
     const fixtureState = useApiStateEntry('fixtures', fixture.id)
@@ -44,12 +45,18 @@ export const FixtureWidget = memoInProduction(
       return null
     }
 
+    const activeMapping = useShallowEqualMemo(
+      mapping
+        ? fixtureType.mapping.filter(channel => mapping.includes(channel))
+        : fixtureType.mapping
+    )
+
     return (
       <FixtureStateWidget
         icon={iconLight}
         title={getFixtureName(fixture, fixtureType)}
         fixtureState={fixtureState}
-        mapping={fixtureType.mapping}
+        mapping={activeMapping}
         {...passThrough}
         onChange={onChange}
         limitedWidth